+++ p4tools/__init__.py
@@ -1,3 +1,3 @@
-__version__ = "0.19.0"
+__version__ = "0.19.1"
 
 from p4tools.io import set_catalog_version, catalog_version